requirements.txt
streamlit==1.39.0
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.13
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2
faiss-cpu==1.8.0
openai==1.3.9
python-dotenv==1.0.0
lxml==4.9.4
validators==0.22.0